Renaming DAO to DTO and fixing camel cases
